+++ src/storage.rs
@@ -13,6 +13,7 @@
 
 pub mod storage_adapter;
 mod storage_handle;
+mod storage_sync;
 pub mod storage_transformer;
 mod storage_value_io;
 pub mod store;
@@ -21,7 +22,6 @@ mod store_prefix;
 
 use std::{path::PathBuf, sync::Arc};
 
-use itertools::Itertools;
 use thiserror::Error;
 
 use crate::{
@@ -34,6 +34,10 @@ use crate::{
 pub use store_key::{StoreKey, StoreKeyError, StoreKeys};
 pub use store_prefix::{StorePrefix, StorePrefixError, StorePrefixes};
 
+pub use self::storage_sync::{
+    ListableStorageTraits, ReadableListableStorageTraits, ReadableStorageTraits,
+    WritableStorageTraits,
+};
 pub use self::storage_transformer::StorageTransformerChain;
 
 pub use self::storage_handle::StorageHandle;
@@ -51,226 +55,6 @@ pub type ListableStorage<'a> = Arc<dyn ListableStorageTraits + 'a>;
 
 /// [`Arc`] wrapped readable and listable storage.
 pub type ReadableListableStorage<'a> = Arc<dyn ReadableListableStorageTraits + 'a>;
-
-/// Readable storage traits.
-pub trait ReadableStorageTraits: Send + Sync {
-    /// Retrieve the value (bytes) associated with a given [`StoreKey`].
-    ///
-    /// Returns [`None`] if the key is not found.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if the store key does not exist or there is an error with the underlying store.
-    fn get(&self, key: &StoreKey) -> Result<MaybeBytes, StorageError>;
-
-    /// Retrieve partial bytes from a list of byte ranges for a store key.
-    ///
-    /// Returns [`None`] if the key is not found.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if there is an underlying storage error.
-    fn get_partial_values_key(
-        &self,
-        key: &StoreKey,
-        byte_ranges: &[ByteRange],
-    ) -> Result<Option<Vec<Vec<u8>>>, StorageError>;
-
-    /// Retrieve partial bytes from a list of [`StoreKeyRange`].
-    ///
-    /// # Arguments
-    /// * `key_ranges`: ordered set of ([`StoreKey`], [`ByteRange`]) pairs. A key may occur multiple times with different ranges.
-    ///
-    /// # Output
-    ///
-    /// A a list of values in the order of the `key_ranges`. It will be [`None`] for missing keys.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if there is an underlying storage error.
-    fn get_partial_values(
-        &self,
-        key_ranges: &[StoreKeyRange],
-    ) -> Result<Vec<MaybeBytes>, StorageError>;
-
-    /// Return the size in bytes of all keys under `prefix`.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `StorageError` if the store does not support size() or there is an underlying error with the store.
-    fn size_prefix(&self, prefix: &StorePrefix) -> Result<u64, StorageError>;
-
-    /// Return the size in bytes of the value at `key`.
-    ///
-    /// Returns [`None`] if the key is not found.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if there is an underlying storage error.
-    fn size_key(&self, key: &StoreKey) -> Result<Option<u64>, StorageError>;
-
-    /// Return the total size in bytes of the storage.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `StorageError` if the store does not support size() or there is an underlying error with the store.
-    fn size(&self) -> Result<u64, StorageError> {
-        self.size_prefix(&StorePrefix::root())
-    }
-
-    /// A utility method with the same input and output as [`get_partial_values`](ReadableStorageTraits::get_partial_values) that internally calls [`get_partial_values_key`](ReadableStorageTraits::get_partial_values_key) with byte ranges grouped by key.
-    ///
-    /// Readable storage can use this function in the implementation of [`get_partial_values`](ReadableStorageTraits::get_partial_values) if that is optimal.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if there is an underlying storage error.
-    fn get_partial_values_batched_by_key(
-        &self,
-        key_ranges: &[StoreKeyRange],
-    ) -> Result<Vec<MaybeBytes>, StorageError> {
-        let mut out: Vec<MaybeBytes> = Vec::with_capacity(key_ranges.len());
-        let mut last_key = None;
-        let mut byte_ranges_key = Vec::new();
-        for key_range in key_ranges {
-            if last_key.is_none() {
-                last_key = Some(&key_range.key);
-            }
-            let last_key_val = last_key.unwrap();
-
-            if key_range.key != *last_key_val {
-                // Found a new key, so do a batched get of the byte ranges of the last key
-                let bytes = (self.get_partial_values_key(last_key.unwrap(), &byte_ranges_key)?)
-                    .map_or_else(
-                        || vec![None; byte_ranges_key.len()],
-                        |partial_values| partial_values.into_iter().map(Some).collect(),
-                    );
-                out.extend(bytes);
-                last_key = Some(&key_range.key);
-                byte_ranges_key.clear();
-            }
-
-            byte_ranges_key.push(key_range.byte_range);
-        }
-
-        if !byte_ranges_key.is_empty() {
-            // Get the byte ranges of the last key
-            let bytes = (self.get_partial_values_key(last_key.unwrap(), &byte_ranges_key)?)
-                .map_or_else(
-                    || vec![None; byte_ranges_key.len()],
-                    |partial_values| partial_values.into_iter().map(Some).collect(),
-                );
-            out.extend(bytes);
-        }
-
-        Ok(out)
-    }
-}
-
-/// Listable storage traits.
-pub trait ListableStorageTraits: Send + Sync {
-    /// Retrieve all [`StoreKeys`] in the store.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if there is an underlying error with the store.
-    fn list(&self) -> Result<StoreKeys, StorageError>;
-
-    /// Retrieve all [`StoreKeys`] with a given [`StorePrefix`].
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if the prefix is not a directory or there is an underlying error with the store.
-    fn list_prefix(&self, prefix: &StorePrefix) -> Result<StoreKeys, StorageError>;
-
-    /// Retrieve all [`StoreKeys`] and [`StorePrefix`] which are direct children of [`StorePrefix`].
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if the prefix is not a directory or there is an underlying error with the store.
-    ///
-    fn list_dir(&self, prefix: &StorePrefix) -> Result<StoreKeysPrefixes, StorageError>;
-}
-
-/// Writable storage traits.
-pub trait WritableStorageTraits: Send + Sync + ReadableStorageTraits {
-    /// Store bytes at a [`StoreKey`].
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] on failure to store.
-    fn set(&self, key: &StoreKey, value: &[u8]) -> Result<(), StorageError>;
-
-    /// Store bytes according to a list of [`StoreKeyStartValue`].
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] on failure to store.
-    fn set_partial_values(
-        &self,
-        key_start_values: &[StoreKeyStartValue],
-    ) -> Result<(), StorageError> {
-        // Group by store key
-        for (key, group) in &key_start_values
-            .iter()
-            .group_by(|key_start_value| &key_start_value.key)
-        {
-            // Read the store key
-            let mut bytes = self.get(key)?.unwrap_or_default();
-
-            // Update the store key
-            for key_start_value in group {
-                let start: usize = key_start_value.start.try_into().unwrap();
-                let end: usize = key_start_value.end().try_into().unwrap();
-                if bytes.len() < end {
-                    bytes.resize(end, 0);
-                }
-                bytes[start..end].copy_from_slice(key_start_value.value);
-            }
-
-            // Write the store key
-            self.set(key, &bytes)?;
-        }
-        Ok(())
-    }
-
-    /// Erase a [`StoreKey`].
-    ///
-    /// Returns true if the key exists and was erased, or false if the key does not exist.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if there is an underlying storage error.
-    fn erase(&self, key: &StoreKey) -> Result<bool, StorageError>;
-
-    /// Erase a list of [`StoreKey`].
-    ///
-    /// Returns true if all keys existed and were erased, or false if any key does not exist.
-    ///
-    /// # Errors
-    ///
-    /// Returns a [`StorageError`] if there is an underlying storage error.
-    fn erase_values(&self, keys: &[StoreKey]) -> Result<bool, StorageError> {
-        let mut all_deleted = true;
-        for key in keys {
-            all_deleted = all_deleted && self.erase(key)?;
-        }
-        Ok(all_deleted)
-    }
-
-    /// Erase all [`StoreKey`] under [`StorePrefix`].
-    ///
-    /// Returns true if the prefix and all its children were removed.
-    ///
-    /// # Errors
-    /// Returns a [`StorageError`] is the prefix is not in the store, or the erase otherwise fails.
-    fn erase_prefix(&self, prefix: &StorePrefix) -> Result<bool, StorageError>;
-}
-
-/// A supertrait of [`ReadableStorageTraits`] and [`ListableStorageTraits`].
-pub trait ReadableListableStorageTraits: ReadableStorageTraits + ListableStorageTraits {}
-
-impl<T> ReadableListableStorageTraits for T where T: ReadableStorageTraits + ListableStorageTraits {}
 
 /// A [`StoreKey`] and [`ByteRange`].
 #[derive(Debug, Clone)]