+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - **Breaking**: Remove `ReadableWritableStorageTraits`, `ReadableWritableStorage`, `ReadableWritableStore`, and `StorageTransformerExtension::create_readable_writable_transformer`
    - These were redundant because `WritableStorageTraits` requires `ReadableStorageTraits` since `6e69a4d`
  - Move sync stores to `storage::store::sync`
+ - Move sync storage traits to `storage_sync.rs`
 
 ## [0.6.0] - 2023-11-16